+++ pkg/clusterrouting/clusterrouting_suite_test.go
@@ -1,0 +1,13 @@
+package clusterrouting
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestClusterRouting(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "ClusterRouting Suite")
+}
